+++ src/main/java/com/bts/personalbudget/controller/fixedbill/FixedBillController.java
@@ -2,7 +2,7 @@ package com.bts.personalbudget.controller.fixedbill;
 
 import com.bts.personalbudget.core.domain.model.FixedBill;
 import com.bts.personalbudget.core.domain.service.fixedbill.FixedBillService;
-import com.bts.personalbudget.core.domain.service.fixedbill.FixedBillServiceFactory;
+import com.bts.personalbudget.core.domain.service.fixedbill.CalcFixedBillFactory;
 import com.bts.personalbudget.mapper.FixedBillMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -17,13 +17,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class FixedBillController {
 
-    private final FixedBillServiceFactory fixedBillServiceFactory;
+    private final FixedBillService fixedBillService;
     private final FixedBillMapper fixedBillMapper;
 
     @PostMapping
     public ResponseEntity<Void> create(@RequestBody FixedBillRequest fixedBillRequest) {
         final FixedBill fixedBill = fixedBillMapper.toModel(fixedBillRequest);
-        final FixedBillService fixedBillService = fixedBillServiceFactory.build(fixedBill.getRecurrenceType());
         fixedBillService.save(fixedBill);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
