+++ src/main/java/com/bts/personalbudget/core/domain/service/fixedbill/CalcFixedBillFactory.java
@@ -1,20 +1,21 @@
 package com.bts.personalbudget.core.domain.service.fixedbill;
 
 import com.bts.personalbudget.core.domain.enumerator.RecurrenceType;
+import com.bts.personalbudget.core.domain.service.fixedbill.calc.CalcFixedBill;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 @Component
-public class FixedBillServiceFactory {
+public class CalcFixedBillFactory {
 
     @Autowired
     private ApplicationContext applicationContext;
 
-    public FixedBillService build(final RecurrenceType recurrenceType) {
+    public CalcFixedBill build(final RecurrenceType recurrenceType) {
         return applicationContext.getBean(
-                FixedBillService.class.getSimpleName() + "." + recurrenceType.name(),
-                FixedBillService.class);
+                CalcFixedBill.class.getSimpleName() + "." + recurrenceType.name(),
+                CalcFixedBill.class);
     }
 
 }
