+++ src/main/java/com/bts/personalbudget/core/domain/service/fixedbill/calc/WeeklyCalcFixedBillImpl.java
@@ -1,22 +1,16 @@
-package com.bts.personalbudget.core.domain.service.fixedbill;
+package com.bts.personalbudget.core.domain.service.fixedbill.calc;
 
-import com.bts.personalbudget.controller.fixedbill.FixedBillRepository;
 import com.bts.personalbudget.core.domain.model.FixedBill;
-import com.bts.personalbudget.mapper.FixedBillMapper;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 @Slf4j
-@Component("FixedBillService.WEEKLY")
-public class WeeklyFixedBillServiceImpl extends FixedBillServiceImpl {
+@Component("CalcFixedBill.WEEKLY")
+public class WeeklyCalcFixedBillImpl implements CalcFixedBill {
 
     private static final int WEEK_DAYS_NUMBER = 7;
-
-    public WeeklyFixedBillServiceImpl(FixedBillRepository fixedBillRepository, FixedBillMapper fixedBillMapper) {
-        super(fixedBillRepository, fixedBillMapper);
-    }
 
     @Override
     public LocalDate calcNextDueDate(final FixedBill fixedBill,
