+++ src/main/java/com/bts/personalbudget/mapper/FinancialMovementMapper.java
@@ -4,10 +4,13 @@ import com.bts.personalbudget.controller.financialmovement.FinancialMovementRequ
 import com.bts.personalbudget.controller.financialmovement.FinancialMovementResponse;
 import com.bts.personalbudget.controller.financialmovement.FinancialMovementUpdateRequest;
 import com.bts.personalbudget.core.domain.entity.FinancialMovementEntity;
+import com.bts.personalbudget.core.domain.enumerator.FinancialMovementStatus;
 import com.bts.personalbudget.core.domain.model.FinancialMovement;
+import com.bts.personalbudget.core.domain.service.installmentbill.InstallmentBill;
 import java.util.List;
 import java.util.UUID;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
 
@@ -24,4 +27,15 @@ public interface FinancialMovementMapper {
     List<FinancialMovementResponse> toResponse(List<FinancialMovement> financialMovement);
     FinancialMovementResponse toResponse(FinancialMovement financialMovement);
 
+    @Mapping(target = "recurrenceBillCode", source = "installmentBill.code")
+    @Mapping(target = "movementDate", source = "installmentBill.purchaseDate")
+    @Mapping(target = "dueDate", source = "installmentBill.nextInstallmentDate")
+    @Mapping(target = "status", source = "financialMovementStatus")
+    FinancialMovement toFinancialMovement(InstallmentBill installmentBill, FinancialMovementStatus financialMovementStatus);
+
+    default List<FinancialMovement> toFinancialMovementList(List<InstallmentBill> installmentBillList, FinancialMovementStatus financialMovementStatus) {
+        return installmentBillList.stream()
+                .map(installmentBill -> toFinancialMovement(installmentBill, financialMovementStatus))
+                .toList();
+    }
 }
