+++ src/main/java/com/bts/personalbudget/mapper/FixedBillMapper.java
@@ -1,10 +1,13 @@
 package com.bts.personalbudget.mapper;
 
-import com.bts.personalbudget.core.domain.entity.FixedBillEntity;
 import com.bts.personalbudget.controller.fixedbill.FixedBillRequest;
+import com.bts.personalbudget.core.domain.entity.CalendarFixedBillEntity;
+import com.bts.personalbudget.core.domain.entity.FixedBillEntity;
 import com.bts.personalbudget.core.domain.model.FixedBill;
 import java.util.List;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Named;
 import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
 
@@ -15,9 +18,17 @@ public interface FixedBillMapper {
 
 
     FixedBill toModel(FixedBillRequest request);
-    FixedBill toModel(FixedBillEntity model);
+    @Mapping(target = "days", source = "calendarFixedBillEntityList", qualifiedByName = "calendarEntityToDays")
+    FixedBill toModel(FixedBillEntity entity);
     List<FixedBill> toModel(List<FixedBillEntity> fixedBillEntities);
     FixedBillEntity toEntity(FixedBill fixedBill);
+
+    @Named("calendarEntityToDays")
+    default List<Integer> calendarEntityToDays(final List<CalendarFixedBillEntity> calendarFixedBillEntityList) {
+        return calendarFixedBillEntityList.stream()
+                .map(CalendarFixedBillEntity::getDayLaunch)
+                .toList();
+    }
 
 
 }
