+++ src/test/java/com/bts/personalbudget/core/domain/service/FinancialMovementServiceTest.java
@@ -3,8 +3,8 @@ package com.bts.personalbudget.core.domain.service;
 import com.bts.personalbudget.core.domain.model.FinancialMovement;
 import com.bts.personalbudget.core.domain.enumerator.FinancialMovementStatus;
 import com.bts.personalbudget.core.domain.enumerator.OperationType;
-import com.bts.personalbudget.core.domain.entity.FinancialMovementFactory;
-import com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty;
+import com.bts.personalbudget.core.domain.factory.FinancialMovementFactory;
+import com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty;
 import com.bts.personalbudget.core.domain.entity.FinancialMovementEntity;
 import com.bts.personalbudget.repository.FinancialMovementRepository;
 import com.bts.personalbudget.mapper.FinancialMovementMapper;
@@ -24,14 +24,14 @@ import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.AMOUNT;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.AMOUNT_PAID;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.DESCRIPTION;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.MOVEMENT_DATE;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.MOVEMENT_DUE_DATE;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.OPERATION_TYPE;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.PAY_DATE;
-import static com.bts.personalbudget.core.domain.entity.FinancialMovementFactory.FinancialMovementProperty.STATUS;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.AMOUNT;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.AMOUNT_PAID;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.DESCRIPTION;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.MOVEMENT_DATE;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.MOVEMENT_DUE_DATE;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.OPERATION_TYPE;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.PAY_DATE;
+import static com.bts.personalbudget.core.domain.factory.FinancialMovementFactory.FinancialMovementProperty.STATUS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
