+++ src/test/java/com/bts/personalbudget/core/domain/service/fixedbill/calc/YearlyFixedBillServiceTest.java
@@ -1,63 +1,23 @@
-package com.bts.personalbudget.core.domain.service;
+package com.bts.personalbudget.core.domain.service.fixedbill.calc;
 
-import com.bts.personalbudget.controller.fixedbill.FixedBillRepository;
-import com.bts.personalbudget.core.domain.entity.CalendarFixedBillEntity;
-import com.bts.personalbudget.core.domain.entity.FixedBillEntity;
-import com.bts.personalbudget.core.domain.entity.FixedBillFactory;
-import com.bts.personalbudget.core.domain.enumerator.FixedBillStatus;
+import com.bts.personalbudget.core.domain.factory.FixedBillFactory;
 import com.bts.personalbudget.core.domain.enumerator.RecurrenceType;
 import com.bts.personalbudget.core.domain.model.FixedBill;
-import com.bts.personalbudget.core.domain.service.fixedbill.YearlyFixedBillServiceImpl;
-import com.bts.personalbudget.mapper.FixedBillMapper;
 import java.time.LocalDate;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 public class YearlyFixedBillServiceTest {
 
     @InjectMocks
-    private YearlyFixedBillServiceImpl yearlyFixedBillService;
-
-    @Mock
-    private FixedBillRepository repository;
-
-    @Spy
-    private FixedBillMapper fixedBillMapper = FixedBillMapper.INSTANCE;
-
-    @Test
-    public void shouldSaveYearlyFixedBill() {
-        List<Integer> days = List.of(1, 365);
-        FixedBill fixedBill = FixedBillFactory.buildModel(RecurrenceType.YEARLY, days);
-        saveTest(fixedBill, days);
-    }
-
-    @Test
-    public void shouldThrowsExceptionWhenDayIsInvalidForYearlyFixedBill() {
-        List<Integer> days = List.of(366);
-        FixedBill fixedBill = FixedBillFactory.buildModel(RecurrenceType.YEARLY, days);
-        assertThrows(RuntimeException.class, () -> yearlyFixedBillService.save(fixedBill));
-    }
-
-    @Test
-    public void shouldThrowsExceptionWhenThereIsEmptyMandatoryFields() {
-        FixedBill fixedBill = FixedBillFactory.buildModel(RecurrenceType.YEARLY, List.of());
-        assertThrows(RuntimeException.class, () -> yearlyFixedBillService.save(fixedBill));
-    }
+    private YearlyCalcFixedBillImpl yearlyFixedBillService;
 
     @Test
     @DisplayName("Próximo data de vencimento - Tipo Anual / Dia 60-bissexto (29/02) / data base de ano bissexto e anterior do vencimento")
@@ -239,35 +199,6 @@ public class YearlyFixedBillServiceTest {
         FixedBill fixedBill = FixedBillFactory.buildModel(recurrenceType, days, leapYear);
         LocalDate nextDueDateOptional = yearlyFixedBillService.calcNextDueDate(fixedBill, baseData);
         assertEquals(expectedDueDate, nextDueDateOptional);
-    }
-
-    private void saveTest(FixedBill fixedBill, List<Integer> days) {
-        yearlyFixedBillService.save(fixedBill);
-
-        ArgumentCaptor<FixedBillEntity> argumentCaptor = ArgumentCaptor.forClass(FixedBillEntity.class);
-
-        verify(repository).save(argumentCaptor.capture());
-
-        FixedBillEntity entity = argumentCaptor.getValue();
-
-        assertNotNull(entity.getCode());
-        assertEquals(fixedBill.getRecurrenceType(), entity.getRecurrenceType());
-        assertEquals(fixedBill.getOperationType(), entity.getOperationType());
-        assertEquals(fixedBill.getDescription(), entity.getDescription());
-        assertEquals(fixedBill.getAmount(), entity.getAmount());
-        assertEquals(fixedBill.getStartDate(), entity.getStartDate());
-        assertEquals(fixedBill.getEndDate(), entity.getEndDate());
-        assertTrue(entity.getFlagActive());
-        assertEquals(FixedBillStatus.ACTIVE, entity.getStatus());
-
-        List<CalendarFixedBillEntity> calendarFixedBillEntityList = entity.getCalendarFixedBillEntityList();
-
-        assertEquals(days.size(), calendarFixedBillEntityList.size());
-        for (int i = 0; i < days.size(); i++) {
-            assertEquals(days.get(i), calendarFixedBillEntityList.get(i).getDayLaunch());
-            assertTrue(calendarFixedBillEntityList.get(i).getFlgActive());
-            assertFalse(calendarFixedBillEntityList.get(i).getFlgLeapYear());
-        }
     }
 
 }
