+++ src/test_integration/java/com/bts/personalbudget/core/domain/repository/FinancialMovementRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.bts.personalbudget.core.domain.repository;
 
 import com.bts.personalbudget.core.domain.entity.FinancialMovementEntity;
-import com.bts.personalbudget.core.domain.entity.FinancialMovementFactory;
+import com.bts.personalbudget.core.domain.factory.FinancialMovementFactory;
 import com.bts.personalbudget.repository.FinancialMovementRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
